requirements-build.txt
pyinstaller>=6.16,<7.0
Pillow>=10.0,<13.0
